feat(app): reorganize and expand tests
Create 2 repositories and test for tag namespacing
