+++ crates/app/tests/helpers/repo.rs
@@ -1,0 +1,64 @@
+use drawbridge_type::{RepositoryConfig, RepositoryName};
+
+use reqwest::StatusCode;
+
+pub async fn create(
+    cl: &reqwest::Client,
+    addr: &str,
+    name: &RepositoryName,
+    conf: RepositoryConfig,
+) {
+    let url = format!("{}/{}", addr, name);
+
+    let res = cl.head(&url).send().await.unwrap();
+    assert_eq!(
+        res.status(),
+        StatusCode::NOT_FOUND,
+        "{}",
+        res.text().await.unwrap()
+    );
+
+    let res = cl.get(&url).send().await.unwrap();
+    assert_eq!(
+        res.status(),
+        StatusCode::NOT_FOUND,
+        "{}",
+        res.text().await.unwrap()
+    );
+
+    let res = cl.put(&url).json(&conf).send().await.unwrap();
+    assert_eq!(
+        res.status(),
+        StatusCode::CREATED,
+        "{}",
+        res.text().await.unwrap()
+    );
+
+    let res = cl.head(&url).send().await.unwrap();
+    assert_eq!(
+        res.status(),
+        StatusCode::OK,
+        "{}",
+        res.text().await.unwrap()
+    );
+    // TODO: Validate headers
+
+    let res = cl.get(&url).send().await.unwrap();
+    assert_eq!(
+        res.status(),
+        StatusCode::OK,
+        "{}",
+        res.text().await.unwrap()
+    );
+    assert_eq!(res.json::<RepositoryConfig>().await.unwrap(), conf);
+    // TODO: Validate headers
+
+    let res = cl.put(&url).json(&conf).send().await.unwrap();
+    assert_eq!(
+        res.status(),
+        // TODO: This should not result in conflict, since payload is the same.
+        StatusCode::CONFLICT,
+        "{}",
+        res.text().await.unwrap()
+    );
+}
